+++ RunManagerHook.cs
@@ -11,7 +11,9 @@ namespace LoomConfig
         [HarmonyPostfix, HarmonyPatch(nameof(RunManager.ChangeLevel))]
         public static void ChangeLevelPostfix(bool _levelFailed)
         {
-            if (_levelFailed) return; // To avoid getting spammed by Imperium if "Disable Game Over" is enabled
+            // To avoid getting spammed by the Arena, or by Imperium if "Disable Game Over" is enabled
+            if (_levelFailed) return;
+            
             LoomConfigRoomProperties.SetLoomProperties();
         }
     }